# docs/openapi-for-user.yaml
openapi: 3.0.0
info:
  title: aeon pet swagger
  description: aeon pet swagger
  version: 1.0.0
...
paths:
...
  /app/users/{userId}/:
    get:
      tags:
        - user
      parameters:
        - schema:
            type: string
          in: path
          name: userId
          required: true
        - schema:
            type: string
          in: header
          name: authorization
          required: false
        - schema:
            type: string
          in: header
          name: x-device-id
          required: false
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  waonCardNumber:
                    type: string
                  email:
                    type: string
                    nullable: true
                  currentStatus:
                    type: string
                  isReceivePushNotification:
                    type: boolean
                required:
                  - id
                  - waonCardNumber
                  - email
                  - currentStatus
                  - isReceivePushNotification
        '400':
          description: リクエストが不正です
          content:
            application/json:
              schema:
                description: リクエストが不正です
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        '401':
          description: トークンが無効です
          content:
            application/json:
              schema:
                description: トークンが無効です
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - トークンが無効です
                      - デバイスが指定されてません
                      - ユーザーとデバイスとの整合性を確認できません
                      - ユーザーIDが不正です
                required:
                  - error
        '404':
          description: ユーザーが見つかりません
          content:
            application/json:
              schema:
                description: ユーザーが見つかりません
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - ユーザーが見つかりません
                required:
                  - error
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                description: サーバーエラー
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Internal Server Error
                required:
                  - error
    post:
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                surName:
                  type: string
                firstName:
                  type: string
                surNameKana:
                  type: string
                firstNameKana:
                  type: string
                birthYear:
                  type: string
                birthMonth:
                  type: string
                birthDay:
                  type: string
                sex:
                  type: string
                  enum:
                    - MAN
                    - WOMAN
                    - NO_ANSWER
                homePostalNo:
                  type: string
                homeAddress1:
                  type: string
                homeAddress2:
                  type: string
                homeAddress3:
                  type: string
                regularStoreList:
                  type: array
                  items:
                    type: string
                mailAddress:
                  type: string
                isReceiveMailMagazine:
                  type: boolean
              required:
                - surName
                - firstName
                - surNameKana
                - firstNameKana
                - birthYear
                - birthMonth
                - birthDay
                - sex
                - homePostalNo
                - homeAddress1
                - homeAddress2
                - homeAddress3
                - regularStoreList
                - mailAddress
                - isReceiveMailMagazine
        required: true
      parameters:
        - schema:
            type: string
          in: path
          name: userId
          required: true
        - schema:
            type: string
          in: header
          name: authorization
          required: false
        - schema:
            type: string
          in: header
          name: x-device-id
          required: false
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties: {}
        '400':
          description: リクエストが不正です
          content:
            application/json:
              schema:
                description: リクエストが不正です
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        '401':
          description: トークンが無効です
          content:
            application/json:
              schema:
                description: トークンが無効です
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - トークンが無効です
                      - デバイスが指定されてません
                      - ユーザーとデバイスとの整合性を確認できません
                      - ユーザーIDが不正です
                required:
                  - error
        '404':
          description: ユーザーが見つかりません
          content:
            application/json:
              schema:
                description: ユーザーが見つかりません
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - ユーザーが見つかりません
                required:
                  - error
        '422':
          description: 既に会員登録済みです
          content:
            application/json:
              schema:
                description: 既に会員登録済みです
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - 既に会員登録済みです
                required:
                  - error
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                description: サーバーエラー
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Internal Server Error
                  errorSubCode:
                    type: number
                  errorSubMessage:
                    type: string
                    enum:
                      - FITSHOP API Error
                required:
                  - error
...